library_api/requirements.txt
asgiref==3.6.0
certifi==2022.12.7
charset-normalizer==3.1.0
Django==4.2
django-extensions==3.2.1
djangorestframework==3.14.0
gunicorn==25.1.0
idna==3.4
packaging==26.0
pillow==12.1.1
pytz==2023.3
requests==2.28.2
sqlparse==0.4.3
tzdata==2025.3
urllib3==1.26.15
whitenoise==6.11.0
